Acquisition automatically stops if data source is lost
Changes to the DataThread, SourceNode, and UIComponent classes allow data
threads to disable callbacks if they are no longer receiving input. So
far this has only been tested with the Intan Board, but it works really well.
Unplugging the board in the middle of acquisition allows the thread to exit,
and callbacks to be disabled, without any seg faults.