Major revisions to FilterList class
The FilterList is now an OpenGLCanvas. It now uses custom fonts (which have
been placed in the Resources/Fonts/ directory), and a completely customized
design. Nevertheless, it seems to retain the same functionality as the old
TreeView-based FilterList.

Minimizing the FilterList is now possible, which allows the DataViewport
to extend the entire length of the window.