Fixed bug in SourceNode
The IntanThread was occasionally looking for data after acquisition was disabled,
which would trigger the SourceNode to attempt to manually disable callbacks.
The only negative effect of this was that if the user was currently viewing
a different signal chain, it would switch back to the Intan chain unexpectedly.
This is no longer the case, as the SourceNode checks to see whether or not
it received a proper disable signal from the ProcessorGraph before notifying
the UI that data may have been dropped.