Fixed bug in LfpDisplayCanvas
When an LfpDisplayCanvas was hidden for more than a few seconds, it would lose
track of the displayBufferIndex, causing it to pause for a few seconds when
it comes back into view. Now, the DataViewport sends a "refresh" reminder
to the canvas, to remind it to update its buffer indices in order to
display new data.