The spikeviewer shows up without segfaults! That was more painful than it should have been.
The SpikeDisplayNode object isn't working properly. Queueing of spikes there appeared to be what was causing the problem.

I need to consider an alternative way of handling the SpikeObjects. In the past I've used a circular buffer but that has
problems with having a set size. I can use an stl::queue but queues are not thread safe, which is something that we
definitely need.

Anyway the plots show up and dance when the viewer is turned on.  I still need to implement a Stereotrode plot and then the
tetrode plot. I'll probably derive them from the ElectrodePlot base class instead of creating a base classs from which all
three plot types are derived from.

Finally I need to get a better handle on the setup of the OpenGLCanvas and how its sized.  How the "Magical" scroll bars work
and how to auto place the plots.

Do we force all the plots to be of the same type? Or do we mix in Electrode Plots with Tetrode Plots?  If so how do we orient them all?