Implemented rudimentary recording functionality for continuous channels
If continuous channels are toggled for recording within their respective editors,
a "writeContinuousBuffer" function will be activated within the RecordNode. This
function doesn't do anything yet, but it will make it easy to add diskwriting
capabilities in the near future. Once a stable format for events has been established,
a similar function will be implemented for writing event buffers to disk.