AudioComponent closes device when not in use. Closes #10.
The AudioComponent now closes the audio device when it's not in use,
saving CPU cycles. Previously, background CPU levels would be between
8 and 10%. Now, when acquisition is not active (and audio device settings
are not being actively edited), background CPU levels drop to 2%. More
optimization will be necessary to get it down to zero, but 2% is still
a big improvement over what we had previously.