Loading Parameters and altering Parameter Buttons based on active channels
Allows parameters to be set from saved XML and changes the
parameterEditor buttons (currently for discrete parameters only) to
reflect which parameters are currently used by selected and
non-selected channels. This fixes a problem where the buttons would
always show what the last press was rather than the current values for
the active channels (e.g. If you changed the parameters on Channel 2,
and then switched to select only Channel 1, the GUI would incorrectly
indicate that the settings for Channel 2 were active). Additionally,
allows the user to see what parameter values are used for channels not
currently selected.