Use processor ID number as flag for saving
Previously, every event would be saved multiple times if
it passed through multiple processors. Now,
GenericProcessor::getNumSamples() checks for existing TTL
events and changes the processor ID (byte 2) to zero.
If the RecordNode encounters an event without a positive
processor ID, it will ignore it.

The only potentially dangerous thing about this is that it
involves modifying the MidiBuffer. But as long as the size
of the buffer doesn't change (just the value of certain
bytes), this should be fine.