basic repaint() restructuring done - still requires refactoring
Basically, we used to have many lfpChannelDisplays in the lfpDisplay,
each with its own repaint() calls. Now, because they all share one big
bitmap display, we need to clear a section of bitmap before we can draw
new data. This has to happen in lfpDisplay::Paint(), not in
lfpChannelDisplay::Paint() because channels overlap and doing it one at
a time would paint over already drawn data form the same buffer.
 Because the repaint() calls are asynchronous, this clearing in the
parent and then drawing the child component isn’t guaranteed to be in
sync - instead, the lfpChannelDisplay pixel-wise plotting is now done
in a new function, and the only redraw() is the one of lfpDisplay.