mitigated vertical artifact bug
though there is still a small display issue if the end of the display
buffer is hit because now only 1-2samples are associated with the pixel
and this causes the min/max and histogram displays to look odd.
shouldn’t be a big issue in real applications but worth fixing when we
inevitably refactor this whole plugin.