Added VisualizerEditor::windowClosed callback
This might be very Cyclops Stimulator specific addition, and there might be a better way to do this:
DataWindow is used only by VisulaizerEditors. So I added a callback to the editor whenever an (open) DataWindow is closed.
I don't know how useful this might be for plugins other than for Cyclops.

Motivation: Since the datawindow is also shared by CyclopsEditors, when a window is closed, only the editor which created the window would get it's windowSelector button toggled back to "off".

With this PR, the button toggles, and the editor's windowClosed callback is called, in which the message is passed to peers through an internally maintained structure -- so that they can toggle their buttons too.