Eliminate offset in audio out channels
The Intan recording module allows users to route two channels to the
audio jack on the acquisition board. Up to this point, selecting channel
1 in the params tab would output software channel 2. In addition,
channel 1 is inaccessible.

This stems from a mismatch between the one-based indexing used by Open
Ephys GUI and the zero-based indexing that the Intan ecosystem uses.