Add support for MS Visual Studio 2015
I have changed the open-ephys.jucer file to include a build
configuration for MS Visual Studio 2015. Additionally, I have added a
post-build command to move the bitfile into the build directory
following compilation.

I was able to compile the main program and load the RHD2000 module, but
was unable to compile the plugins and therefore could not confirm that
it was working correctly.