Make datestring in .continuous file headers fixed width
As discussed in #129, the date_created field of .continuous file headers does
not have fixed width days of the month, hours, or seconds, which makes parsing
the header difficult and often ambiguous. I have changed the RecordNode class
to produce fixed width datestrings, which should fix this problem.

I was unable to compile with these changes, but the previous commit (226bfab)
won't compile for me either. However, I can make these changes to an earlier
commit and compile successfully.