Fix bug with data channel and audio monitor DataChannel objects are copied when they are prepped for acquisition. Also, the copy constructor for DataChannel did not inherit the recording, enabled, or monitored state from the original when it is copied.
Changing the state during acquisition resulted in changing the copy as desired, but not updating the original. Stopping acquisition and resuming would not be up to date with user configuration. Changing state before enabling acquisition would not propagate to the copy, because the copy constructor hardcoded the value. Both of these issues were changed in this commit.

Since RecordNodes are handled in the same way, it is likely that this bug exists there too. However that behavior has not been changed yet.