Added continuous "on" button to pulse pal plugin canvas
I added a "continuous" button to each output channel on the pulse pal plugin canvas so that users can set the pulse pal's "continuous" triggering state to either "on" or "off".

I added this because when the user fires up the open ephys pulse pal plugin, all of the options predict a finite pulse train.  If the "continuous" state has previously been set to "on" another way, the open ephys plugin options cannot turn it off, and the user may wonder why their pulse train continues indefinitely.  The continuous on/off trigger state is also nice to have for debugging experimental setups.