Fix nullptr dereferencing when plugin not found
Don't call setEnabledState() in the editor's constructor because it
forwards to the processor's setEnabledState() which in turn calls the
editor's enable(), but editor is still nullptr at that point since
it's being constructed.
Doesn't change the initial intent of creating the plugin in a disabled
state but fixes crash when loading configurations which refer to
non-existing plugins for instance.