Fix Phase Detector trigger channel always returning 0
Phase Detector was setting trigger channel to 0 despite the channel selector interface showing correct channel as selected. Changed the way how `param` value was fetched. Before it was simply converting the `var` value to `int` and then using that as local index. Now, it fetches the `param` value as an `array`, then gets the first value of that `array` (since max selecetable channels is 1) and uses that as local index.