Implementing RecordNodeTests along with additional test infrastructure
Implementing the tests for the RecordNode, which necessitated some
changes to the underlying test harnesses / infrastructure. Also, a few
bugs in the RecordNode itself were fixed, as they were causing crashes
(and presumingly can occur during real runtime too).

Main part of the infrastructure change is found in `ProcessorTester`.
Aside from making it a class that can be composed instead of inherited,
it now spins up more of the required OpenEphys infrastructure, including
the MessageCenter, ControlPanel, and ProcessorGraph. It automatically
adds the `FakeSourceNode` and then allows for creating a new processor
and attaching it to the end of this chain.

The immediate reason for the change was because `RecordNode` expects a
decent amount of global state to exist, such as those states in
AccessClass, but presumingly this change can be applied for other
plugins.