Fix crash in DataViewport on reload
When a processor with a Visualizer fails to load, a placeholder processor is created. Adding tab for the placeholder processor results in a crash.
This fix adds a isVisualizerEditor virtual method to GenericEditor which returns false by default, and true if it is a VisualierEditor. This allows the DataViewport to verify befor calling addTab.